code/requirements.txt
Python==3.5.3
Flask==0.12.1

# python packages needed
sqlite3
flask
picamera
RPi.GPIO
adafruit-pca9685
